Create mithril-gui.py
First iteration starting off as a qt based gui that is designed for kde plasma in the breeze dark theme which facilitates gocryptfs management in an intuitive and simplified way, but also has a toggle for an integrated drop down terminal. Assumes that fish is preinstalled as a terminal option. Terminal allows interaction via command line interface and viewing the commands and processes in real time. GUI has a settings menu with additional advanced options toggles, options for setting default preferences, as well as an alternative advanced display mode to switch to a full featured GUI rather than the default simplified GUI. Also included; the ability to create different profiles for the default GUI so that you can have different toggles and default preferences and save them to a profile that you can select as your default GUI. Written with Arch systems in mind.